Update patch set 1
Patch Set 1:

> Can you upstream the `.d.ts` file maybe?
> 
> Also: why is the `.d.ts` file required in the first place? Can it be prevented to have the file in the first place? E.g. if you can port the files in your fork to ESM you probably don't need the `.d.ts`.
> 
> If none of these options are applicable, I would suggest putting it in legacy-defs.d.ts for now at the top. Hopefully that minimizes the odds of you running into merge conflicts. However, keep in mind that we have no knowledge of the files you have in your fork and it will be difficult for us to know how not to create unnecessary merge conflicts.

Since the Root module is at the bottom of the dependency graph, it can't list "common" in its ts_library deps.

I've tried it with both a separate file and appended at the end of legacy-defs.d.ts. Both work just fine, and that might be the route we go. 

As to upstreaming: We'll discuss internally. I'm kind of loathe to do it because the definitions refer to infra that we use for error reporting and tracking. I think the "right answer" is to actually probably break this into a separate module entirely, although we want to catch errors starting at initialization, so the outcome will probably be something hybrid. (This is a stopgap to port stuff back into Runtime.js which had been removed).

Thanks for helping out with this.

Patch-set: 1