Update patch set 2
Patch Set 2: Code-Review+1

It is difficult to see what the patch is. For future relands, please click the "Revert" button on the revert of your CL. That then creates a new CL. You locally patch that in your repository with `git cl patch <CL_NUMBER_HERE>`. Then you make the changes to reland and then you `git cl upload`. This then allows the reviewers to see the diff in the patchsets and inspect it.

For now, I suspect that the logic of deleting the property has an additional waitfor to deflake it? That LGTM.

Patch-set: 2
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1