Reformat all module.json files
This patch makes it so that all module.json files use the same
consistent indentation that matches the rest of the codebase.

Note that this patch does not "roll CodeMirror", although a
pre-commit hook forces me to include that phrase in the commit
message because the module.json file in the relevant folder is
modified.

Change-Id: Ib31ed5232461e2f80bf05f31105fc919f6639b0a