Show unbound breakpoints as disabled
We need the UI to distinguish between a successfully bound breakpoint
and one that has not been bound. This change makes unbound breakpoints
look like disabled breakpoints, though that could be changed to a
different icon in the future.

This change also eliminates a race condition in some tests by waiting
for a breakpoint to be bound after setting it.

Fixed: chromium:1063864, chromium:1064581
Change-Id: Id181a19efd7f9939a1555368b9367aaa22af0fd2