Update patch set 5
Patch Set 5:

(1 comment)

This may be out-of-scope for this particular change but when testing using Narrator, only the first time I added classes was announced. Subsequent changes did not result in an announcement. Stepping through the code confirmed you were still calling UI.ARIAUtils.alert with the correct values, so I suspect this may be an existing bug with that helper as opposed to your changes.

In particular I suspect in UI.ARIAUtils.alert the indirect reference using aria-describedby to the element whose label is being updated is causing Narrator to miss the update. However this seems to be in place to mitigate an NVDA bug according to nearby comments, so I'd edit this with caution. Might be worth leaving as-is for now and logging a separate bug (possibly against Narrator).

What do you think?

Patch-set: 5
Reviewer: Gerrit User 1390564 <1390564@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>