Update patch set 6
Patch Set 6: Code-Review+1

To the best of my knowledge this seems to be a reasonable fix for the problem. Is there any documentation on this part of the front-end, especially the threat model and the relevant parts in the code for extensions?

Patch-set: 6
Reviewer: Gerrit User 1183894 <1183894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1