Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 5:
> 
> (1 comment)
> 
> This may be out-of-scope for this particular change but when testing using Narrator, only the first time I added classes was announced. Subsequent changes did not result in an announcement. Stepping through the code confirmed you were still calling UI.ARIAUtils.alert with the correct values, so I suspect this may be an existing bug with that helper as opposed to your changes.
> 
> In particular I suspect in UI.ARIAUtils.alert the indirect reference using aria-describedby to the element whose label is being updated is causing Narrator to miss the update. However this seems to be in place to mitigate an NVDA bug according to nearby comments, so I'd edit this with caution. Might be worth leaving as-is for now and logging a separate bug (possibly against Narrator).
> 
> What do you think?

Yikes.  Well, I am the one who wrote the alert function that way to fix its functionality with NVDA.  It's unfortunate that Narrator doesn't consume it the same way.

I'll file a bug on myself to figure out the NVDA/Narrator behavior with alert.  If writing the alert function without the described-by workaround works for Narrator, I think we code it that way and file a bug against NVDA.  Either way, I agree that the alert change is out of scope for this CL.

Patch-set: 6