Update patch set 2
Patch Set 2:

This is kind of cool how the tooltip glides around, but it feels inconsistent with the other toolbar items, e.g. if I use the Performance panel and hover over the '(no recordings)' area next to the screenshots checkbox, the tooltip doesn't move even when I move my cursor around inside the element. I guess this is because it's actually a styled button, but to the user it might seem weird that it behaves differently to the checkboxes/labels.

Is it possible to make the tooltip stay still once displayed? I don't think the horizontal position is all that important. One cool thing with the tooltips on the buttons is that the vertical position is always the exact same which is less distracting.

Basically I'm proposing that it behaves like the ('no recording') element, where the Y position is always the same regardless of mouse position inside the element, and the X position doesn't change while the tooltip is shown. I don't mind if the original X position is influenced by where the first mouse-over occurred or not.

WDYT?

Patch-set: 2
Reviewer: Gerrit User 1167239 <1167239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>