Update patch set 3
Patch Set 3:

> I have added a change to apply the same behavior as buttons have to checkboxes. I think it's not a good solution long-term because the Tooltip component determines how the tooltips work for the entire application. It might be better to expose a new tooltip option for changing the behavior. Also, I don't know if the behavior that the tooltip depends on the mouse position is actually desired in any cases.
> 
> With this change, the reported components behave as expected but I noticed that on the Network tab it's not consistent yet (the throttling dropdown has a different behavior).

Agreed, it's hard to know if any caller relies on the current behavior. Do you think it will be difficult to add an option to let the caller decide? We can keep the old behavior in the default case, and leave a TODO in the code to check the other callers later. It looks like we need to plumb the option through e.g. Toolbar in Toolbar.js so if it's too much hassle then let's go with your current solution and leave a TODO in the code.

Patch-set: 3
Reviewer: Gerrit User 1167239 <1167239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>