Reposition tooltip if the mouse moves
The related bug describes a situation when the tooltip looks misaligned
if you navigate to it from another opened toolip. It happens because
the new tooltip is opened on the first mouse event and is placed
relative to the border of the anchor element (where the mouse is).
Therefore, it looks misaligned because the user moves the mouse
pointer further. This change implements the repositioning of the tooltip
 on subsequent events.  An alternative approach would be to position the
tooltip relative to the anchor element only and avoid using mouse
pointer coordinates.

Bug: 1039631
Change-Id: Idfcb59318ce9c44f8397573d759e17dcdd08314f