Set background color for scrollbars in dark mode
The color of the scrollbar track was transparent in the dark mode.
It works for most of the panels but it does not work well for the
live expressions panel in Console. This change sets the scrollbar
background color to rgb(36, 36, 36) in dark mode. This color is
the background color for most of the UI elements with a scrollbar.

Bug: 953677
Change-Id: Ib382b38cd6c16852bd7c0240d99b25abe73b110c