Count WTF-8 byte size without encoding
In `ExpandableTextPropertyValue`, we’re doing a full WTF-8 encode
just to get the byte size of the result. Instead, we can compute
the total byte size without the overhead of doing any encoding.
For the string `'Iñtërnâtiônàlizætiøn☃💩'.repeat(1_000_000)` this
reduces the runtime cost from 166 ms to 98 ms. For the string
`'ASCII-only'.repeat(1_000_000)` the runtime cost increases only
slightly, from 42 ms to 45 ms.

This patch introduces `StringUtilities.countWtf8Bytes` to do exactly
that. Note that WTF-8 [1] is like UTF-8 with additional support for
lone surrogates. We need to support lone surrogates in this case,
since `ExpandableTextPropertyValue` is potentially dealing with
JavaScript strings.

[1]: https://simonsapin.github.io/wtf-8/

Bug: chromium:1024721
Change-Id: I00d0cd8ecd328daddb43b2476229d5e4e47038e8