Count WTF-8 byte size without encoding
In `ExpandableTextPropertyValue`, we’re doing a full WTF-8 encode
just to get the byte size of the result. Instead, we can compute
the total byte size without the overhead of doing any encoding.

This patch introduces `StringUtilities.countWtf8Bytes` to do exactly
that. Note that WTF-8 [1] is like UTF-8 with additional support for
lone surrogates. We need to support lone surrogates in this case,
since `ExpandableTextPropertyValue` is potentially dealing with
JavaScript strings.

[1]: https://simonsapin.github.io/wtf-8/

Change-Id: I00d0cd8ecd328daddb43b2476229d5e4e47038e8