Update patch set 8
Patch Set 8:

I understand the security requirement here, but since there is a way to activate extensions on chrome:// URLs and such, that way should also apply to Developer Tools extensions.
The way is to enable chrome://flags/#extensions-on-chrome-urls
Can you pass it to the front end and disable/enable extensions according to it so it would still be possible?

I imagine debugging the Developer Tools themselves (for the team) could be easier with certain extensions, so it is in your interest to support that debugging option.

Patch-set: 8
CC: Gerrit User 1104195 <1104195@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>