Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Tim,
> 
> Philip helped me today to solve a timing issue that I had with my e2e test.
> Using "await addBreakpointForLine" would not actually wait for the content to load,
> so we added a check to wait for that particular line. 
> 
> This is probably not enough though, it would be good to wait for content before proceeding.
> PTAL :)  
> 
> Kim

Could you upload your CL so that I can take a look? The `openFileInEditor` function in `sources-helpers` is supposed to wait for the file to be loaded. We should (also) fix the logic there if it turns out to be incorrect.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>