[issues] Add issue counter to console counters in top menu bar
This CL adds the issue counter to the top menu bar. It relies on the
issue model being enables from the start. It uses the main target to
get the issues model, and then sticks with that. This works, because
we bubble all issues up to the local root (i.e. all arrive in the main
target model).

Once we add issues for workers, etc. we might have to revise this.

Bug: chromium:1057979
Change-Id: I1421d5d2243c5ad28b1e66b9f9e7739c310192bd