Update Rollup to 2.3.3
I have checked https://github.com/rollup/rollup/releases/tag/v2.0.0
and none of the breaking changes appear to be applicable to us.

I have manually confirmed that with this build on a production bundle
of DevTools, the output is okay and the build is working as intended.

R=jacktfranklin@chromium.org
DISABLE_THIRD_PARTY_CHECK=Update rollup

Bug: 1068145
Change-Id: I6a7ce6eaacb29fdfd1462642c976b81ce9c20e03