Add infobar about issues panel to console
This CL adds a sticky infobar to the console. It shows up whenever
issues are detected and notifies users about the existence of the new
issues panel. The bar disappears automatically when the user navigates
to a page for which no issues are detected.

Screenshot: https://imgur.com/Jk95x7n

Fixed: chromium:1061876
Change-Id: I16d47f2a620f33bd52a80f1c3005cfe107c29e38