Use typescript-eslint parser for localizability pipeline
While clang-format and eslint support the modern javascript features,
the localizability pipeline did not. It was using esprima, which is
a parser that is not well maintained and does not support modern
features.

For our ESLint configuration, we are using the typescript-eslint
parser, which is compatible with espree, the parser powering ESLint.
Both espree and typescript-eslint have an API that is very similar
to esprima.

Therefore, update the localizability pipeline to use the espree API
with the concrete implementation of the API by typescript-eslint.

In the future, the localizability pipeline should probably be an
ESLint plugin so that we can unify the parsing experience.

A follow-up CL will remove all usages of esprima.

R=jacktfranklin@chromium.org

Fixed: 1068966
Change-Id: Idcffb8d649f006d7cf0b3de0ee0d886cb3848230