Network: Expand Parent Frame Nodes on Reveal
This CL fixes a bug where clicking on a search result in the network
tool would not properly reveal nodes if they were in a collapsed
group frame node.

Viewport datagrids do not actually attempt to reveal nodes that are not
attached on the root node flatlist - such as when a
parent node is collapsed. This causes newtwork nodes being revealed to
check if their parent is a frame node, and expand if so, allowing the
datagrid to properly reveal and select the node.

This bug originated when separating selection from opening the sidepane,
as attempting to open the sidepane on a frame node child would cause the
parent frame to expand. Which effectively "hid" this behavior until
selection no longer opened the sidepane.

See the bug description for good repro steps/site.

Bug: 1068747
Change-Id: Ie491a8f82286376fa50dd9922b16c44d6737c0a1