Determine whether request params should be decoded by default
Not all requests have parameters URL-encoded. This change checks
if a request is application/x-www-form-urlencoded and decodes the
parameters by default only in this case (as suggested by phistuck in
the related bug). Additionally, improved the update code to avoid
loosing the scroll position when the view is toggled.

Fixed: 1022088
Change-Id: Ic238541de9986ffd5b2a431991c9ca494a6afe30