Update patch set 1
Patch Set 1:

How it looks when not hovering (this CL did not change this behaviour): 
https://imgur.com/s6oJGpZ

BEFORE: how it looked when hovering on border (note the very poor contrast):
https://imgur.com/20WPNda

AFTER: we swap to a dark font when hovering border which improves the contrast:
https://imgur.com/NvfiGkj

This fixes part of this bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1053978 with regards to the border and hovering. There is another bug around showing the `position` attribute which I'm working on right now.

Patch-set: 1