Fix Devices Settings regression and other ListWidgets
A recent CL that added keyboard navigation to ListWidget didn't include
focus styles for most ListWidgets, which resulted in a loss of visible
focus  (screencasts in bug). This CL adds focus styles for all
ListWidgets and makes the checkboxes in the Request Blocking and Devices
Settings pane operable by pressing space on their list items.

Bug: 1068065
Change-Id: Ic317db7712c9b558b74308eb4aaf4c4be89a5068