Network: Update Search to Activate Requests when clicked
This CL is a follow up to
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2141147

This restores the old search behavior, which would select and show the
side panel with the appropriate tab/line number. This behavior
was broken when separating request selection and activation in the
network panel.

Demo: https://imgur.com/BSHgcVH

*Note that the grid properly selects nodes, but may not scroll them into view.
This is due to an existing issue where having the sidepanel
open can prevent nodes from scrolling into view. I tested against v79, and
this behavior was present there as well.

Bug: 1068747
Change-Id: I7bab79fc4ad8d57c062a0fe69696bf2623ed6d13