Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Similar feedback about respondWith -- I'd like to see the ServiceWorker-related calls be conceptually represented as a descendant of the "Waiting" section. LMK your thoughts. Thanks!

Are all ServiceWorker-related timings conceptually a subset of the time spent "Waiting", though? At least the linked image suggests the completion of the fetch may extend beyond the waiting state. That makes sense to me as it seems feasible some data could be returned from the worker before the full fetch has completed. If that's the case I don't think representing these as descendants would be accurate.

Patch-set: 1
CC: Gerrit User 1390564 <1390564@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>