Update patch set 2
Patch Set 2:

Hi everyone, 
This CL contains a proposed usage for Lighthouse i18n in the devtools. (More details on the Description). This is not intended to be checked in, it is only to clarify the API shape. When the i18n library has been checked in, a different CL containing the actual changes will follow up.


PTAL.

Patch-set: 2
CC: Gerrit User 1312148 <1312148@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1378407 <1378407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1341658 <1341658@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1319038 <1319038@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false