Update patch set 1
Patch Set 1:

I feel like an e2e-test is most suitable here. You probably want to craft a network response that creates the unreachable URL scenario. I think the pattern in `network-datagrid.ts` is appropriate here. E.g. you write a couple of `.rawresponse` files with the HTTP headers you are looking for. Then you request these files and assert that (when going to the network panel) the cookies are shown.

For the LightHouse issue, I think it is a bit harder. We have no helpers yet for LightHouse and I fear that is more involved. Therefore I think it is better to disregard these for now in tests. We probably should come up with a clear testing strategy for the LightHouse integration.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>