Add worker spawn test and clean up pages folder
This adds a new test that verifies a breakpoint set in a script used
in workers will be hit if a new worker is spawned after the breakpoint
is set. Also moved stuff out of the resources/pages folder that belongs
elsewhere. Also fixed the reformatting of the minified file so that
its source map makes sense.

Bug: 1062308
Change-Id: Ifa0a31513332d423ef4529abf6f357b9d238dd42