Update patch set 3
Patch Set 3:

PTAL when you get the chance. 

One note: I noticed cookies stopped appearing under affected resources after the rebase, but I see the same issue in today's canary so it seems unrelated to my changes. I debugged a bit and it looks like the structure of events from the backend related to cookies is different than expected - not sure if this is already known or not.

Patch-set: 3