Hide frame highlight when the mouse leaves Console context dropdown
When a frame is selected in Console context dropdown, it's highlighted
using highlightFrame CDP method. The highlightFrame handler is using
a different tool to highlight the frame compared to the normal case.
This highlight tool renders a slightly different box for the iframe
and it does not switch to other elements when you hover over them.

The separation of highlighting methods into different tools was
probably introduced to simplify the code and, perhaps, support
other use cases. So instead of changing how highlight tools are
working in the backend, the suggestion is to hideHighlight for
the highlighted frame as soon as the user moves the mouse outside
of the dropdown which should suffice to resolve reported issue.

Fixed: 999440
Change-Id: Ib490ec5cc274b1404d0489c31a2edef185de3cf2