[issues] Rework how issues are associated with affected resources
Resources like network requests or cookies are currently associated via
a private property to a list of various issues. This has some drawbacks,
mainly a timing dependency that the connection between resources -> issue
has to be made when the issue is received from the backend.

This CL changes that by getting rid of this connection altogether and
doing a lookup each time we want to check if a given resource (like
network request) has any issues associated with them. A follow up CL
will introduce some supporting data structures that should elevate
most of the performance problems introduced by this CL (for huge
issue counts).

Bug: chromium:1070577
Change-Id: I61fa60e8861b35719f66c9e12e67cdc1896f49c9