Update patch set 4
Patch Set 4:

Hi Peter!

Sorry for the looong delay. This is a small change that I did for fixit, but 
I tried quite a long time to create a Karma test for this but failed 
(also with Jack's help; we decided that it would be too much work to test this properly),
and didn't get to upload it until today. Would you mind having a look?
Too late for the badge, but this should be uploaded anyway.

This change imho only fixes a symptom, but I believe that the root cause
lies somewhere else. In any case, the calculation should not result in
a negative sum nevertheless.

If you want to talk about what's happening in the code I can give you more insight
since I had a look at it (ping me, or VC). An example where things go wrong are:
1. event1, start: 0, end: 5
2. event2, start: 1, end: 6

Since I do not know the timeline too well, from what I can see in the code
is that 1. should be a parent to 2. and should not end before 2. So the
example should only be hypothetical, but it happens for the trace that was
attached in the bug report. Therefore I am saying that I think it's more a symptom
than the root cause. I can investigate this further during the next fixit.

Patch-set: 4
Reviewer: Gerrit User 1167239 <1167239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1396836 <1396836@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false