Update patch set 13
Patch Set 13:

(3 comments)

> Patch Set 13:
> 
> (3 comments)
> 
> So a couple things I would like to understand a bit more:
> 
> 1. Do we support parallelization still? (It seems not, but I may have missed it)

We do not, as we are not using it atm. In a follow-up CL I will test with `mocha-parallel-tests`.

> 2. Do we support running in slowmo or with timeouts between steps (again, seems like not)

Will add that back. I didn't want to make the initial CL as big and this felt as a candidate for a follow-up. I will integrate it in this one.

> I would like particularly the second one to be in place before LGTM. The parallelization could come afterwards, but again I'd like to bring it in as quickly as we can.

Patch-set: 13
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>