[issues] Refactor issue displaying
Use the abstract class issue for displaying instead of AggregateIssue.
This decouples the displaying from the issue defintions.

This CL also moves IssueCategory to Issue, and improves issue lookup
by introducing a map from requestId to a list of requests in the
network log. The list is necessary because the back-end reuses
request ids occasionally (e.g. for redirects).

Bug: chromium:1072304
Change-Id: I9cf8ee962ecf9e19628992b13a538a2a947ed3f4