[issues] Refactor issue displaying
Use the abstract class issue for displaying instead of AggregateIssue.
This decouples the displaying from the issue defintions.

This CL moves IssueCategory to Issue, and speeds up issue lookup
in NetworkLog to improve performance of chromium:1070577 which is now
used more often. Performance is improved by introducing a map from
requestId to a list of requests in the network log. The list is
necessary because the back-end reuses request ids occasionally
(e.g. for redirects and in some caching situations).

Bug: chromium:1072304, chromium:1070577
Change-Id: I9cf8ee962ecf9e19628992b13a538a2a947ed3f4