Reattach the target before reloading
Some not yet resolved backend problem breaks the usual flow of events
after the anchor element navigation happens within the DevTools
(e.g., <a href="#anchor" />). One of the issues is that styles events
are missing after the DevTools's reloaded. This change reattaches the
target before reloading DevTools to mitigate the issue.

Fixed: 972001
Change-Id: Ie815a9a7e5dda8d39686068c8064e19bc31285b7