Centralize shortcut titles
There are a few parts of the UI that assume that shortcuts will only be
one key (e.g. the shortcut tooltips displayed for some buttons), so this
CL updates them to access a new KeyboardShortcut.title() method to get
the name of a shortcut rather than just getting the name of a single
keypress from its Descriptor. The current/legacy Shortcuts settings tab
was one of these places, but I left it alone because it will be
obsoleted by the new custom shortcuts tab in the future and in the
meantime the only users who will have any chords enabled in the DevTools
will be those who have opted in to the custom shortcuts experiment and
enabled VS Code shortcuts.

Custom shortcuts doc: https://docs.google.com/document/d/1oOPSWPxCHvMoBZ0Fw9jwFZt6gP4lrsrsl8DEAp-Hy7o/edit#

Bug: 174309
Change-Id: Ic56fb14129a417ed441fc46401af226db52c8905