Properly cleanup on process teardown when pre-emptively stopping mocha
If developers would kill the Mocha process before completion, it would
not run the `after` hook. This meant that its subprocesses (most notably
the Puppeteer process) would not properly cleanup.

To mitigate that, listen for the process exit and `SIGINT` signal (e.g.
a developer doing `ctrl + c`) and cleanup. However, if the `after` hook
ran succesfully, we shouldn't be cleaning up twice.

R=aerotwist@chromium.org

Bug: 1071369
Change-Id: I072b2b662309af1c508a7877c4b897cf915bed71