Update ListControl to explicitly indicate ARIA list positions
This CL improves the accessibility of the Command Menu by fixing an
issue where screen reader software (NVDA) will read aloud incorrect
list positions for each item (e.g. "5 of 24").

This is caused by usages of the ListControl component where elements
in the list are loaded into the DOM on-demand rather than all at once.
The screen reader cannot infer that there are more items in the list
because they are not present in the DOM.

This CL updates the ListControl to use two ARIA properties to
explicitly indicate the position of each item, and the total number
of items in the list.

 - aria-setsize : explicitly indicates the size of the list,
                  useful when only a portion of the list may be
                  present in the document [1]

 - aria-posinset : explicitly indicates the position of an item
                   in a list [2]

ListControl is used elsewhere in DevTools, including the call stack
list and soft dropdown menus. The explicitly defined list positions
should continue to be accurate in those use cases and should not cause
a regression in behavior, which I have tested in addition to the
command menu.

Before: https://i.imgur.com/O0Tm1Wx.gif

After: https://i.imgur.com/Ma0CxKf.gif

[1] https://w3c.github.io/aria/#aria-setsize
[2] https://w3c.github.io/aria/#aria-posinset


Bug: 1073151
Change-Id: I635f3526dbde38e001a6ab375a45b4e933b9a554