Issues with affected network requests can choose the network tab
This CL allows different issue types that reference network requests,
to choose which tab in the network panel the link navigates to.
For example, SameSite cookie issues should navigate to the "Cookies"
tab rather then the "Headers" tab.

The logic for this is entirely implemented on the UI side, rather
than polluting SDK classes with types from the Network view layer.

R=sigurds@chromium.org

Bug: chromium:1057979
Change-Id: Ia494daad892b482183933465c6e032d1f8a5949c