[issues] Properly de-duplicate requests in AggregatedIssue
Previously the AggregedIssue class used a {Set<string>} for affected
network requests. This would implicitly de-duplicate requests with
the same requestId. An architecture change turning this into a
{Set<AffectedRequest>} got rid of the implicit de-duplication,
resulting in the same request showing multiple times in the UI.

This CL fixes this by re-introducing a map where we can de-duplicate
on the requestId.

Drive-by: Extract {StubIssue} into its own file, as its used by two
different unit tests now.

R=sigurds@chromium.org

Bug: chromium:1068116
Change-Id: I47ed91c7a3f8d123026edbeb96e81c5b73adac28