Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > Implementation LGTM. However, I would propose waiting with submitting until all pre-existing violations are removed. Then we can be sure the rule works as expected. If we don't, we might find out that when we actually try to do so there are problems and require follow-up fixes as well.
> 
> I did that locally and confirmed that I get an error on `assert.equal` that is fixed by swapping to `deepEqual` or `strictEqual`. I was just trying to keep this CL small. Given that, are you happy to submit?

Yeah this CL is good as-is, but I would say that you first do a separate cleanup CL and get that in, before submitting this one (with the rule also turned on). In the end it is your call.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>